Tests/LibWeb/Text/input/localStorage.html

<script src="include.js"></script>
<script>
    test(() => {

        println(localStorage["test"])
        println(localStorage.getItem("test"))

        localStorage["test"] = "value";
        println(localStorage["test"]);
        println(localStorage.getItem("test"));

        localStorage.setItem("test", "other");
        println(localStorage["test"]);
        println(localStorage.getItem("test"));

        // Ensure indexed properties work
        println(localStorage[0])
        println(localStorage.getItem(0))

        localStorage[0] = "foo";
        println(localStorage[0]);
        println(localStorage.getItem("0"));
    });
</script>
