Tests/LibWeb/Text/expected/localStorage.txt

undefined
null
value
value
other
other
undefined
null
foo
foo
